+++ packages/shared/package.json
@@ -22,7 +22,6 @@
         "zod": "catalog:zod"
     },
     "devDependencies": {
-        "@types/cookie": "^1.0.0",
         "bun-types": "^1.3.8",
         "drizzle-kit": "catalog:drizzle",
         "tsdown": "^0.20.1",
@@ -38,7 +37,9 @@
         "db:sync:push": "drizzle-kit push --config=drizzle.sync.config.ts",
         "db:sync:studio": "drizzle-kit studio --config=drizzle.sync.config.ts",
         "build": "tsdown",
-        "dev": "tsdown --watch --no-clean"
+        "dev": "tsdown --watch --no-clean",
+        "check-types": "pnpm exec tsc -p tsconfig.json --noEmit",
+        "lint": "pnpm exec eslint ."
     },
     "exports": {
         ".": "./dist/index.js",
